# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JAC Rental Booking Service API",
        "description": "Standalone microservice handling all rental property booking operations for the JAC platform. Authentication uses Bearer tokens issued by the JAC monolith (Laravel Passport).",
        "contact": {
            "email": "dev@jac.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "L5_SWAGGER_CONST_HOST",
            "description": "Current Environment"
        }
    ],
    "paths": {
        "/api/v1/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "List user bookings",
                "description": "Returns all bookings belonging to the authenticated user, newest first.",
                "operationId": "daf1d30d180ba251d3a594b8cc06fcc5",
                "responses": {
                    "200": {
                        "description": "List of bookings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookingSummary"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/bookings/{id}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get booking details",
                "description": "Returns full details for a specific booking owned by the authenticated user.",
                "operationId": "e510ab8fc8a07c21f4f5aa7d77eb2cca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/bookings/estimate": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Estimate booking price",
                "description": "Calculates the total price for a unit over selected dates including per-night breakdown. Does NOT create a booking or hold. Use this when the user is browsing dates before committing.",
                "operationId": "00a6b4c4a418da02a4f9d0827a8e5c6b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "unit_id",
                                    "check_in_date",
                                    "check_out_date"
                                ],
                                "properties": {
                                    "unit_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "check_in_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-09-01"
                                    },
                                    "check_out_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-09-04"
                                    },
                                    "guests_count": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "currency": {
                                        "description": "SAR | USD | YER_S | YER_N",
                                        "type": "string",
                                        "example": "SAR"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Price estimate breakdown",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PriceEstimate"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid date range"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/bookings/initiate-payment": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Initiate the payment flow",
                "description": "Creates a 15-minute date hold on the unit and returns a payment URL. The booking record is only created AFTER the payment gateway confirms payment via the webhook. Call this when the user taps \"Pay Now\".",
                "operationId": "44c51ba1b12c14e74bb1c4d34c21e0b2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "unit_id",
                                    "check_in_date",
                                    "check_out_date"
                                ],
                                "properties": {
                                    "unit_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "check_in_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-09-01"
                                    },
                                    "check_out_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-09-04"
                                    },
                                    "guests_count": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "SAR"
                                    },
                                    "guest_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "guest_phone": {
                                        "type": "string",
                                        "example": "+966501234567"
                                    },
                                    "guest_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "guest_note": {
                                        "type": "string",
                                        "example": "Late check-in please"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment initiation details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "hold_token": {
                                            "type": "string",
                                            "example": "HOLD-ABCDE12345XY"
                                        },
                                        "expires_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "total_amount": {
                                            "type": "number",
                                            "example": 450
                                        },
                                        "currency": {
                                            "type": "string",
                                            "example": "SAR"
                                        },
                                        "payment_url": {
                                            "type": "string",
                                            "example": "https://api.example.com/api/v1/payments/initiate?hold_token=HOLD-ABC&amount=450"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/bookings/{id}/cancel": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Cancel a booking",
                "description": "Cancels a confirmed booking owned by the authenticated user. Push notification is sent to the user via the monolith bridge.",
                "operationId": "d9764a5646618d284ae2c1f31ece8eb2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Change of plans"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking cancelled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Booking cancelled successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Cannot cancel booking in current state"
                    },
                    "404": {
                        "description": "Booking not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BookingDetail": {
                "description": "Full booking details",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 88
                    },
                    "reference_no": {
                        "type": "string",
                        "example": "BK-ABCD1234"
                    },
                    "status": {
                        "type": "string",
                        "example": "confirmed"
                    },
                    "payment_status": {
                        "type": "string",
                        "example": "paid"
                    },
                    "check_in_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-09-01"
                    },
                    "check_out_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-09-04"
                    },
                    "nights_count": {
                        "type": "integer",
                        "example": 3
                    },
                    "guests_count": {
                        "type": "integer",
                        "example": 2
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 450
                    },
                    "currency": {
                        "type": "string",
                        "example": "SAR"
                    },
                    "guest_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "guest_phone": {
                        "type": "string",
                        "example": "+966501234567"
                    },
                    "guest_email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "guest_note": {
                        "type": "string",
                        "nullable": true
                    },
                    "property": {
                        "$ref": "#/components/schemas/PropertySummary"
                    },
                    "unit": {
                        "$ref": "#/components/schemas/UnitSummary"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "BookingSummary": {
                "description": "Compact booking representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 88
                    },
                    "reference_no": {
                        "type": "string",
                        "example": "BK-ABCD1234"
                    },
                    "status": {
                        "type": "string",
                        "example": "confirmed"
                    },
                    "payment_status": {
                        "type": "string",
                        "example": "paid"
                    },
                    "check_in_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-09-01"
                    },
                    "check_out_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-09-04"
                    },
                    "nights_count": {
                        "type": "integer",
                        "example": 3
                    },
                    "guests_count": {
                        "type": "integer",
                        "example": 2
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 450
                    },
                    "currency": {
                        "type": "string",
                        "example": "SAR"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "PriceEstimate": {
                "description": "Price breakdown and total for booking",
                "properties": {
                    "unit_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "check_in_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-09-01"
                    },
                    "check_out_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-09-04"
                    },
                    "nights_count": {
                        "type": "integer",
                        "example": 3
                    },
                    "unit_price": {
                        "type": "number",
                        "format": "float",
                        "example": 150
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 450
                    },
                    "currency": {
                        "type": "string",
                        "example": "SAR"
                    },
                    "nightly_breakdown": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "date": {
                                    "type": "string",
                                    "format": "date",
                                    "example": "2026-09-01"
                                },
                                "price": {
                                    "type": "number",
                                    "format": "float",
                                    "example": 150
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "PropertySummary": {
                "description": "Compact property representation returned in list views",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Sea View Villa"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "city_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "type_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "avg_rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.5
                    },
                    "rating_count": {
                        "type": "integer",
                        "example": 12
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": true
                    },
                    "star_rating": {
                        "type": "integer",
                        "example": 4
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "logo": {
                        "type": "string",
                        "example": "properties/villa1.jpg",
                        "nullable": true
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Passport Bearer token issued by the JAC monolith. Obtain via the monolith /api/v1/auth/login endpoint.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}